Agrega detalles visuales

+++ src/app/education-programs/components/list/list.component.ts
@@ -1,8 +1,10 @@
 import { Component } from '@angular/core';
 import { FormControl, Validators } from '@angular/forms';
+import { MatDialog } from '@angular/material/dialog';
 import { MatTableDataSource } from '@angular/material/table';
 import { Router, ActivatedRoute } from '@angular/router';
 import { ToastrService } from 'ngx-toastr';
+import { AcceptDialog } from 'src/app/shared/dialogs';
 import { EducationProgramService } from '../../services';
 
 @Component({
@@ -29,6 +31,7 @@ export class ListComponent{
   constructor(
     private router: Router,
     private toastSerice: ToastrService,
+    public dialog: MatDialog,
     private activatedroute: ActivatedRoute,
     private educationProgramService: EducationProgramService
   ) {
@@ -56,12 +59,18 @@ export class ListComponent{
   }
 
   delete(element) {
-    this.educationProgramService.delete(element.id).subscribe(result => {
-      if (result.success) {
-        this.removeElement(element);
-        this.toastSerice.success('Removido');
+    const ref = this.dialog.open(AcceptDialog);
+    ref.afterClosed().toPromise().then(result => {
+      if (result) {
+        this.educationProgramService.delete(element.id).subscribe(result => {
+          if (result.success) {
+            this.removeElement(element);
+            this.toastSerice.success('Removido');
+          }
+        })
       }
-    })
+    });
+
   }
 
   private addElement(element) {
